excise the codebase from some C99 format specifiers. (%zu -> %lu)
git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/trunk@7120 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/multiplay.c
@@ -1893,7 +1893,7 @@ const char* getPlayerColourName(unsigned int player)
 		N_("Cyan")
 	};
 
-	ASSERT(player < ARRAY_SIZE(playerColors), "player number (%d) exceeds maximum (%zu)", player, ARRAY_SIZE(playerColors));
+	ASSERT(player < ARRAY_SIZE(playerColors), "player number (%d) exceeds maximum (%lu)", player, (unsigned long) ARRAY_SIZE(playerColors));
 
 	if (player >= ARRAY_SIZE(playerColors))
 	{
